+++ app/src/main/java/com/example/habittrackerapp/repository/HabitRepository.kt
@@ -1,0 +1,41 @@
+package com.example.habittrackerapp.repository
+
+import android.content.Context
+import android.widget.Toast
+import androidx.navigation.NavHostController
+import com.google.firebase.firestore.ktx.firestore
+import com.google.firebase.ktx.Firebase
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
+
+class HabitRepository {
+    val db = Firebase.firestore
+    fun addHabit(
+        name: String,
+        description: String,
+        repeat: List<String>,
+        reminder: LocalTime,
+        startFrom: LocalDate,
+        context: Context,
+        navController: NavHostController,
+    ) {
+
+        val habit = hashMapOf(
+            "name" to name,
+            "description" to description,
+            "repeat" to repeat,
+            "reminder" to reminder,
+            "startFrom" to startFrom.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
+        )
+        db.collection("habits").add(habit).addOnSuccessListener {
+            Toast.makeText(context, "Habit added successfully", Toast.LENGTH_SHORT).show()
+            navController.navigate("/home")
+        }.addOnFailureListener {
+            Toast.makeText(context, "Something Went Wrong", Toast.LENGTH_SHORT).show()
+        }
+    }
+
+
+
+}
